merge

+++ Symulator/Matrices.cs
@@ -11,10 +11,10 @@ namespace Symulator
 {
     public static class Matrices
     {
-        static int[,] distance;  //meters
+        static long[,] distance;  //meters
         static long[,] time;     // seconds
 
-        public static int[,] Distance
+        public static long[,] Distance
         {
             get
             {
@@ -45,7 +45,7 @@ namespace Symulator
             if (PackagesList.numberOfPackages > 0)
             {
                 int infinity = -1;//tymczasowo -1,  mozna zmienic -> -1 = brak czasu -> brak trasy
-                distance = new int[PackagesList.numberOfPackages + 1, PackagesList.numberOfPackages + 1];
+                distance = new long[PackagesList.numberOfPackages + 1, PackagesList.numberOfPackages + 1];
                 time = new long[PackagesList.numberOfPackages + 1, PackagesList.numberOfPackages + 1];
 
 
@@ -204,7 +204,7 @@ namespace Symulator
 
             if (File.Exists(filename))
             {
-                distance = new int[PackagesList.numberOfPackages + 1, PackagesList.numberOfPackages + 1];
+                distance = new long[PackagesList.numberOfPackages + 1, PackagesList.numberOfPackages + 1];
 
                 using (StreamReader sr = new StreamReader(filename))
                 {
